% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avif.R
\name{read_avif}
\alias{read_avif}
\title{Read an AVIF image}
\usage{
read_avif(source, ..., ptype = raw(), normalize = FALSE)
}
\arguments{
\item{source}{A file path or a raw vector.}

\item{ptype}{Prototype such as raw(), 0L or 0.0 that defines the output type.}

\item{normalize}{If \code{TRUE}, output a normalized (0-1) array.}
}
\value{
An RGB array.
}
\description{
Read an AVIF image from a file or a raw vector into an RGB array.
}
\examples{
read_avif("8bpc.avif")
read_avif("10bpc.avif", ptype = 0L)
read_avif(readBin("12bpc.avif", "raw", file.size("12bpc.avif")), ptype = 0., normalize = TRUE)
}
