% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avif.R
\name{write_avif}
\alias{write_avif}
\title{Write an AVIF Image}
\usage{
write_avif(
  image,
  target = NULL,
  ...,
  speed = 6L,
  quality = 60L,
  alpha_quality = 60L,
  format = 444L,
  codec = NULL,
  jobs = NULL
)
}
\arguments{
\item{image}{A raw or integer vector with dimensions (height, width, channel).}

\item{target}{A file path, or \code{NULL} for a raw vector.}

\item{...}{Unused.}

\item{speed}{Encoder speed in [0, 10] where 0 is the slowest, 10 is the fastest.}

\item{quality}{Color quality in [0, 100] where 100 is lossless.}

\item{alpha_quality}{Alpha quality in [0, 100] where 100 is lossless.}

\item{format}{YUV format, must be one of 444, 422, 420 or 400.}

\item{codec}{Codec for encoding, automatic if \code{NULL}.}

\item{jobs}{Number of encoder threads, must be greater than 0, or \code{NULL} for all cores.}
}
\value{
\code{NULL} if \code{target} is a file path, otherwise a raw vector.
}
\description{
Create an AVIF image from an RGB array into a file or return a raw vector.
}
\examples{
rgb_array <- hcl.colors(700) |>
  col2rgb() |>
  as.raw() |>
  array(c(3, 700, 100)) |>
  aperm()
save_path <- file.path(tempdir(), "8bpc.avif")
write_avif(rgb_array, save_path, speed = 0L, quality = 100L)
writeBin(write_avif(rgb_array), save_path)
}
