% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avif.R
\name{write_avif}
\alias{write_avif}
\title{Write an AVIF image}
\usage{
write_avif(image, target = NULL)
}
\arguments{
\item{image}{A raw vector with dimensions (height, width, channel).}

\item{target}{A file path, or \code{NULL} for a raw vector.}
}
\value{
\code{NULL} if \code{target} is a file path, or a raw vector.
}
\description{
Create an AVIF image from an RGB array into a file or return a raw vector.
}
\examples{
rgb_array <- hcl.colors(700) |>
  col2rgb() |>
  as.raw() |>
  array(c(3, 700, 100)) |>
  aperm()
write_avif(rgb_array, "8bpc.avif")
writeBin(write_avif(rgb_array), "8bpc.avif")
}
